Fix handling of symbolic links on Linux
When a file is a symbolic link to some other file the proper behavior
when calculating the checksum of a package is to use the link path
instead of the content of the pointed file.

Additionally the `getPackageFileListWithoutVcs` procedure is fixed to
return also the symbolic link files.

Related to nim-lang/nimble#127